(feat) Add outdated docs warning
Add a warning on top of the documentation content if you're looking at
V1 docs. It will include a notice and a link to the latest version of
the docs. In cases where the exact docs do not exist, this will land the
user on the more generic topic page.